+++ changelog.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2024-10-20
+
+### 🐛 Bug Fixes
+
+- *(cli)* Version detection
+
+### ⚙️ Miscellaneous Tasks
+
+- *(cli)* Table width fixed considering table width
+- *(cli)* Updated version detection
+- *(cli)* Updated get versio to call update check
+- *(changelog)* Version update
+- *(cli)* Prod build updated
+- *(cli!)* Removed axios and refactored some parts
+- *(repo)* Removed sql and npmignore update
+- *(cli)* Removed axios and refactored some parts
+
 ## [1.1.4] - 2024-10-19
 
 ### 🚀 Features
@@ -19,6 +36,7 @@ All notable changes to this project will be documented in this file.
 ### ⚙️ Miscellaneous Tasks
 
 - *(cli)* Width fix
+- *(changelog)* Updated version minor
 
 ## [1.1.3] - 2024-10-15